index.html index.css
The basic structure of the home page is complete, and optimized the results of the previous code, using the design ideas of the version of the heart, respectively, with floats and positioning to complete the layout of some of the functions on the page. We will add some other pictures later, but for the time being, we have used the same picture.